publish GalaCredit platform

+++ backend/app/models/ecard_pool.py
@@ -1,6 +1,6 @@
 from datetime import datetime
 
-from sqlalchemy import Column, DateTime, Float, Integer, String
+from sqlalchemy import Column, DateTime, Float, Integer, String, Numeric
 
 from app.core.database import Base
 
@@ -11,7 +11,7 @@ class EcardPool(Base):
     id = Column(Integer, primary_key=True, index=True)
     account = Column(String(100), nullable=False, unique=True, index=True)
     password = Column(String(100), nullable=False)
-    face_value = Column(Float, default=0.0, nullable=False, index=True)
+    face_value = Column(Numeric(18, 2), default=0.0, nullable=False, index=True)
     expires_at = Column(DateTime, nullable=False, index=True)
     status = Column(String(20), nullable=False, default="AVAILABLE", index=True)
     loan_id = Column(Integer, nullable=True, index=True)
