+++ backend/app/models/loan_installment.py
@@ -1,6 +1,6 @@
 from datetime import datetime
 
-from sqlalchemy import Column, DateTime, Float, Integer, String
+from sqlalchemy import Column, DateTime, Float, Integer, String, Numeric
 from sqlalchemy.orm import relationship
 
 from app.core.database import Base
@@ -15,20 +15,20 @@ class LoanInstallment(Base):
     due_date = Column(DateTime, nullable=False, index=True)
     status = Column(String(20), default="PENDING", nullable=False, index=True)
 
-    principal_amount = Column(Float, default=0.0, nullable=False)
-    interest_amount = Column(Float, default=0.0, nullable=False)
-    guarantee_fee_amount = Column(Float, default=0.0, nullable=False)
-    due_amount = Column(Float, default=0.0, nullable=False)
+    principal_amount = Column(Numeric(18, 2), default=0.0, nullable=False)
+    interest_amount = Column(Numeric(18, 2), default=0.0, nullable=False)
+    guarantee_fee_amount = Column(Numeric(18, 2), default=0.0, nullable=False)
+    due_amount = Column(Numeric(18, 2), default=0.0, nullable=False)
 
-    paid_principal_amount = Column(Float, default=0.0, nullable=False)
-    paid_interest_amount = Column(Float, default=0.0, nullable=False)
-    paid_guarantee_fee_amount = Column(Float, default=0.0, nullable=False)
-    paid_amount = Column(Float, default=0.0, nullable=False)
+    paid_principal_amount = Column(Numeric(18, 2), default=0.0, nullable=False)
+    paid_interest_amount = Column(Numeric(18, 2), default=0.0, nullable=False)
+    paid_guarantee_fee_amount = Column(Numeric(18, 2), default=0.0, nullable=False)
+    paid_amount = Column(Numeric(18, 2), default=0.0, nullable=False)
 
-    reduced_principal_amount = Column(Float, default=0.0, nullable=False)
-    reduced_interest_amount = Column(Float, default=0.0, nullable=False)
-    reduced_guarantee_fee_amount = Column(Float, default=0.0, nullable=False)
-    reduction_amount = Column(Float, default=0.0, nullable=False)
+    reduced_principal_amount = Column(Numeric(18, 2), default=0.0, nullable=False)
+    reduced_interest_amount = Column(Numeric(18, 2), default=0.0, nullable=False)
+    reduced_guarantee_fee_amount = Column(Numeric(18, 2), default=0.0, nullable=False)
+    reduction_amount = Column(Numeric(18, 2), default=0.0, nullable=False)
 
     settled_at = Column(DateTime, nullable=True)
     created_at = Column(DateTime, default=datetime.now, nullable=False)
