+++ backend/app/models/momo_transaction.py
@@ -1,6 +1,6 @@
 from datetime import datetime
 
-from sqlalchemy import Column, DateTime, Float, Integer, String, Text
+from sqlalchemy import Column, DateTime, Float, Integer, String, Text, Numeric
 
 from app.core.database import Base
 
@@ -18,7 +18,7 @@ class MomoTransaction(Base):
     provider_reference = Column(String(120), nullable=True, unique=True, index=True, comment="服务商流水号")
     idempotency_key = Column(String(120), nullable=False, unique=True, index=True, comment="幂等键")
     phone = Column(String(20), nullable=False, comment="MoMo手机号")
-    amount = Column(Float, nullable=False, default=0, comment="交易金额")
+    amount = Column(Numeric(18, 2), nullable=False, default=0, comment="交易金额")
     status = Column(String(24), nullable=False, default="PENDING", index=True, comment="交易状态")
     request_payload = Column(Text, nullable=True, comment="请求报文快照")
     response_payload = Column(Text, nullable=True, comment="服务商响应报文快照")
